\name{AF2}
\alias{A2F}
\title{Map of the Relative Accuracy of EU-derived PGS for African Ancestry}
\description{

The data frame provides a Map of the Relative Accuracy of cross-ancestry Polygenic Score Prediction for European-derived genomic scores when used to predict phenotypes of individuals of African Ancestry.



}

\keyword{datasets}
