\name{getSegmentsChr}
\alias{getSegmentsChr}
\title{Computes small chromosome segments within one chromosome.}
\usage{
  getSegmentsChr(bp, minBPSize = 10e3, minSize = 10, firstSegment = 1)
}
\arguments{
    \item{bp}{
    }
    \item{minBPSize}{
    }
    \item{minSize}{
    }
    \item{firstSegment}{
    }
}
