\name{MAP_UKB}
\alias{MAP_UKB}
\title{A map of the Relative Accuracy of EU-derived PGS for non-EU Ancestry Groups}
\description{
  This data frame provides a map of the Relative Accuracy of cross-ancestry Polygenic Score Prediction for European-derived 
  genomic scores when used to predict phenotypes of individuals of African, Caribbean, East Asian, or South Asian Ancestry.
}

\keyword{datasets}
