\docType{package}
\name{MCANOVA-package}
\alias{MCANOVA-package}
\title{A Package to Estimate Local Relative Accuracy and PGS portability.}
\description{
  - Two functions, getSegments() and getSegmentsChr(), to estimate small chromosome segments/windows.
  - A function to run the MC-ANOVA relative accuracy estimation, MC_ANOVA().
  - An interactive shiny app, PGS_portability_app(), that provides SNP information and relative accuracy estimates for user-supplied inputs.
}
