\name{getSegments}
\alias{getSegments}
\title{Computes small chromosome segments; calls getSegmentsChr.R}
\usage{
  getSegments(bp, chr, minBPSize = 10e3, minSize = 10, firstSegment = 1, verbose = FALSE)
}
\arguments{
    \item{bp}{
    }
    \item{chr}{
    }
    \item{minBPSize}{
    }
    \item{minSize}{
    }
    \item{firstSegment}{
    }
    \item{verbose}{
    }
}
\value{
    A numeric vector of segment labels corresponding to each marker.
}
