\docType{package}
\name{MCANOVA-package}
\alias{MCANOVA-package}
\title{A Package to Estimate Local Relative Accuracy and PGS portability.}
\description{
  \itemize{
    \item Two functions, getSegments() and getSegmentsChr(), to estimate small chromosome segments/windows.
    \item A function to run the MC-ANOVA relative accuracy estimation, MC_ANOVA().
    \item An interactive shiny app, PGS_portability_app(), that provides SNP information and relative accuracy estimates for user-supplied inputs.
  }
}
